split types and reduce imports
